refactored look-up-wallet-address to be an standalone application
