+++ src/application/extension-popup-menu/index.ts
@@ -1,2 +1,1 @@
 export { App as ExtensionPopupMenu } from './app';
-export { COMMAND_MAP as EXTENSION_MENU_COMMAND_MAP } from './commands';
