--- src/application/extension-popup-menu/look-up-wallet-address/index.ts
@@ -1,1 +1,0 @@
-export { App as LookUpWalletAddress } from './app';
