+++ src/application/look-up-wallet-address/commands/index.ts
