+++ src/application/look-up-wallet-address/index.ts
@@ -1,0 +1,2 @@
+export { App as LookUpWalletAddress } from './app';
+export { COMMAND_MAP as EXTENSION_MENU_COMMAND_MAP } from './commands';
