+++ src/final/final.tsx
@@ -1,6 +1,7 @@
 import { PolymarketApp } from 'application/polymarket';
 import { useExtensionSettings } from 'shared/extension';
 import { ErrorBoundary } from 'shared/observability';
+import { LookUpWalletAddress } from 'application/look-up-wallet-address';
 
 import { Proposals, UserWidgets } from './widgets';
 
@@ -13,6 +14,7 @@ export const Final = () => {
 
   return (
     <ErrorBoundary>
+      <LookUpWalletAddress />
       <UserWidgets />
       <PolymarketApp />
       <Proposals />
